[#93] Splitting JSON tests to reflect node file structure
