+++ src/FluentSerializer.Json.Tests/Tests/DataNodes/Nodes/JsonArrayTests.Writing.cs
@@ -1,0 +1,58 @@
+using FluentSerializer.Core.TestUtils.Extensions;
+
+using Xunit;
+
+using static FluentSerializer.Json.JsonBuilder;
+
+namespace FluentSerializer.Json.Tests.Tests.DataNodes.Nodes;
+
+public sealed partial class JsonArrayTests
+{
+	[Fact,
+		Trait("Category", "UnitTest"),	Trait("DataFormat", "JSON")]
+	public void AppendTo_HasValue_FormatWriteNull_ReturnsValue()
+	{
+		// Arrange
+		var input = Array(Object());
+		var expected = "[\n\t{\n\t}\n]";
+
+		// Act
+		input.AppendTo(ref _textWriter, true, 0, true);
+		var result = _textWriter.ToString();
+
+		// Assert
+		result.ShouldBeBinaryEquatableTo(expected);
+	}
+
+	[Fact,
+		Trait("Category", "UnitTest"),	Trait("DataFormat", "JSON")]
+	public void AppendTo_HasNoValue_FormatWriteNull_ReturnsEmptyArray()
+	{
+		// Arrange
+		var input = Array();
+		var expected = "[\n]";
+
+		// Act
+		input.AppendTo(ref _textWriter, true, 0, true);
+		var result = _textWriter.ToString();
+
+		// Assert
+		result.ShouldBeBinaryEquatableTo(expected);
+	}
+
+	[Fact,
+		Trait("Category", "UnitTest"),	Trait("DataFormat", "JSON")]
+	public void AppendTo_HasNoValue_FormatDontWriteNull_ReturnsEmptyArray()
+	{
+		// Arrange
+		var input = Array();
+		var expected = "[\n]";
+
+		// Act
+		input.AppendTo(ref _textWriter, true, 0, false);
+		var result = _textWriter.ToString();
+
+		// Assert
+		result.ShouldBeBinaryEquatableTo(expected);
+	}
+}
