+++ src/FluentSerializer.Json.Tests/Tests/DataNodes/Nodes/JsonObjectTests.Writing.cs
@@ -1,0 +1,57 @@
+using FluentSerializer.Core.TestUtils.Extensions;
+
+using Xunit;
+
+using static FluentSerializer.Json.JsonBuilder;
+
+namespace FluentSerializer.Json.Tests.Tests.DataNodes.Nodes;
+public sealed partial class JsonObjectTests
+{
+	[Fact,
+		Trait("Category", "UnitTest"),	Trait("DataFormat", "JSON")]
+	public void AppendTo_HasValue_FormatWriteNull_ReturnsValue()
+	{
+		// Arrange
+		var input = Object(JsonProperty);
+		var expected = $"{{\n\t{JsonProperty}\n}}";
+
+		// Act
+		input.AppendTo(ref _textWriter, true, 0, true);
+		var result = _textWriter.ToString();
+
+		// Assert
+		result.ShouldBeBinaryEquatableTo(expected);
+	}
+
+	[Fact,
+		Trait("Category", "UnitTest"),	Trait("DataFormat", "JSON")]
+	public void AppendTo_HasNoValue_FormatWriteNull_ReturnsEmptyObject()
+	{
+		// Arrange
+		var input = Object();
+		var expected = $"{{\n}}";
+
+		// Act
+		input.AppendTo(ref _textWriter, true, 0, true);
+		var result = _textWriter.ToString();
+
+		// Assert
+		result.ShouldBeBinaryEquatableTo(expected);
+	}
+
+	[Fact,
+		Trait("Category", "UnitTest"),	Trait("DataFormat", "JSON")]
+	public void AppendTo_HasNoValue_FormatDontWriteNull_ReturnsEmptyObject()
+	{
+		// Arrange
+		var input = Object();
+		var expected = $"{{\n}}";
+
+		// Act
+		input.AppendTo(ref _textWriter, true, 0, false);
+		var result = _textWriter.ToString();
+
+		// Assert
+		result.ShouldBeBinaryEquatableTo(expected);
+	}
+}
