+++ src/FluentSerializer.Json.Tests/Tests/DataNodes/Nodes/JsonObjectTests.cs
@@ -1,20 +1,11 @@
-using FluentAssertions;
-
-using FluentSerializer.Core.Constants;
-using FluentSerializer.Core.TestUtils.Extensions;
 using FluentSerializer.Core.TestUtils.Helpers;
 using FluentSerializer.Core.Text;
 using FluentSerializer.Json.DataNodes;
-using FluentSerializer.Json.DataNodes.Nodes;
-
-using System;
-
-using Xunit;
 
 using static FluentSerializer.Json.JsonBuilder;
 
 namespace FluentSerializer.Json.Tests.Tests.DataNodes.Nodes;
-public sealed class JsonObjectTests
+public sealed partial class JsonObjectTests
 {
 	private static readonly IJsonValue JsonValue = Value("69");
 	private static readonly IJsonProperty JsonProperty = Property("value", in JsonValue);
@@ -25,128 +16,4 @@ public sealed class JsonObjectTests
 	{
 		_textWriter = TestStringBuilderPool.CreateSingleInstance();
 	}
-
-	#region Parse
-	[Fact,
-		Trait("Category", "UnitTest"),	Trait("DataFormat", "JSON")]
-	public void ParseJson_Valid_ReturnsObject()
-	{
-		// Arrange
-		var expected = Object(JsonProperty);
-		var input = $"{{ {JsonProperty} }}";
-
-		// Act
-		var offset = 0;
-		var result = new JsonObject(input, ref offset);
-
-		// Assert
-		result.Should().BeEquatableTo(expected);
-	}
-
-	[Theory,
-		Trait("Category", "UnitTest"),	Trait("DataFormat", "JSON"),
-		InlineData(""), InlineData(" "), InlineData("  "), InlineData("\t"),
-		InlineData(LineEndings.LineFeed), InlineData(LineEndings.CarriageReturn),
-		InlineData(LineEndings.ReturnLineFeed)]
-	public void ParseJson_ValidWithWhiteSpace_ReturnsObject(string space)
-	{
-		// Arrange
-		var expected = Object(JsonProperty);
-		var expectedEmpty = Object();
-		var input1 = $"{{{space}{JsonProperty}{space}}}{space}";
-		var input2 = $"{{{space}}}{space}";
-
-		// Act
-		var offset1 = 0;
-		var result1 = new JsonObject(input1, ref offset1);
-		var offset2 = 0;
-		var result2 = new JsonObject(input2, ref offset2);
-
-		// Assert
-		result1.Should().BeEquatableTo(expected);
-		result2.Should().BeEquatableTo(expectedEmpty);
-	}
-
-	[Fact,
-		Trait("Category", "UnitTest"),	Trait("DataFormat", "JSON")]
-	public void ParseJson_Empty_ReturnsObject()
-	{
-		// Arrange
-		var expected = Object();
-		var input = "{}";
-
-		// Act
-		var offset = 0;
-		var result = new JsonObject(input, ref offset);
-
-		// Assert
-		result.Should().BeEquatableTo(expected);
-	}
-
-	[Fact,
-		Trait("Category", "UnitTest"),	Trait("DataFormat", "JSON")]
-	public void ParseJson_Incomplete_Throws()
-	{
-		// Arrange
-		var input = "{ \"prop\": 14, ";
-
-		// Act
-		var offset = 0;
-		var result = () => new JsonObject(input, ref offset);
-
-		// Assert
-		result.Should()
-			.ThrowExactly<ArgumentOutOfRangeException>();
-	}
-	#endregion
-
-	#region ToString
-	[Fact,
-		Trait("Category", "UnitTest"),	Trait("DataFormat", "JSON")]
-	public void AppendTo_HasValue_FormatWriteNull_ReturnsValue()
-	{
-		// Arrange
-		var input = Object(JsonProperty);
-		var expected = $"{{\n\t{JsonProperty}\n}}";
-
-		// Act
-		input.AppendTo(ref _textWriter, true, 0, true);
-		var result = _textWriter.ToString();
-
-		// Assert
-		result.ShouldBeBinaryEquatableTo(expected);
-	}
-
-	[Fact,
-		Trait("Category", "UnitTest"),	Trait("DataFormat", "JSON")]
-	public void AppendTo_HasNoValue_FormatWriteNull_ReturnsEmptyObject()
-	{
-		// Arrange
-		var input = Object();
-		var expected = $"{{\n}}";
-
-		// Act
-		input.AppendTo(ref _textWriter, true, 0, true);
-		var result = _textWriter.ToString();
-
-		// Assert
-		result.ShouldBeBinaryEquatableTo(expected);
-	}
-
-	[Fact,
-		Trait("Category", "UnitTest"),	Trait("DataFormat", "JSON")]
-	public void AppendTo_HasNoValue_FormatDontWriteNull_ReturnsEmptyObject()
-	{
-		// Arrange
-		var input = Object();
-		var expected = $"{{\n}}";
-
-		// Act
-		input.AppendTo(ref _textWriter, true, 0, false);
-		var result = _textWriter.ToString();
-
-		// Assert
-		result.ShouldBeBinaryEquatableTo(expected);
-	}
-	#endregion
 }
