+++ src/FluentSerializer.Json.Tests/Tests/DataNodes/Nodes/JsonValueTests.Writing.cs
@@ -1,0 +1,61 @@
+using FluentSerializer.Core.TestUtils.Extensions;
+
+using Xunit;
+
+using static FluentSerializer.Json.JsonBuilder;
+
+namespace FluentSerializer.Json.Tests.Tests.DataNodes.Nodes;
+public sealed partial class JsonValueTests
+{
+	[Fact,
+		Trait("Category", "UnitTest"),	Trait("DataFormat", "JSON")]
+	public void AppendTo_HasValue_FormatWriteNull_ReturnsValue()
+	{
+		// Arrange
+		var input = Value("true");
+		var expected = "true";
+
+		// Act
+		input.AppendTo(ref _textWriter, true, 0, true);
+		var result = _textWriter.ToString();
+
+		// Assert
+		result.ShouldBeBinaryEquatableTo(expected);
+	}
+
+	[Fact,
+		Trait("Category", "UnitTest"),	Trait("DataFormat", "JSON")]
+	public void AppendTo_HasNoValue_FormatWriteNull_ReturnsNullValue()
+	{
+		// Arrange
+		var input = Value(null);
+		var expected = "null";
+
+		// Act
+		input.AppendTo(ref _textWriter, true, 0, true);
+		var result = _textWriter.ToString();
+
+		// Assert
+		result.ShouldBeBinaryEquatableTo(expected);
+	}
+
+	/// <summary>
+	/// Since the property is responsible for hiding itself, values have no support
+	/// for this distinction
+	/// </summary>
+	[Fact,
+		Trait("Category", "UnitTest"),	Trait("DataFormat", "JSON")]
+	public void AppendTo_HasNoValue_FormatDontWriteNull_ReturnsNullValue()
+	{
+		// Arrange
+		var input = Value(null);
+		var expected = "null";
+
+		// Act
+		input.AppendTo(ref _textWriter, true, 0, false);
+		var result = _textWriter.ToString();
+
+		// Assert
+		result.ShouldBeBinaryEquatableTo(expected);
+	}
+}
